Major refactoring; reformatting; flake8 linting
