+++ cleanup-old-gists
@@ -4,22 +4,23 @@ Delete uploaded logfiles from Gist that are older than 30 days.
 
 """
 
-import os, sys
+import os
+import sys
+import requests
+import dateutil.parser
+from datetime import datetime
 from socket import setdefaulttimeout
+
 setdefaulttimeout(120)
 
-import json
-import requests
-from datetime import datetime
-import dateutil.parser
 
 # https://docs.github.com/en/rest/reference/gists
-headers = {'Accept': 'application/vnd.github.v3+json',
-          'Authorization': 'token %s' % os.environ['GITHUBTOKEN']
+headers = {
+    "Accept": "application/vnd.github.v3+json",
+    "Authorization": "token %s" % os.environ["GITHUBTOKEN"],
 }
 
-r = requests.get('https://api.github.com/gists', 
-              headers = headers)
+r = requests.get("https://api.github.com/gists", headers=headers)
 
 
 response_json = r.json()
@@ -31,23 +32,22 @@ gist_delete_limit = 5
 gists_deleted = 0
 for gist in response_json:
     if gists_deleted > gist_delete_limit:
-        print('gist_delete_limit reached - exit for now.')
+        print("gist_delete_limit reached - exit for now.")
         sys.exit(0)
-    if 'id' not in gist:
+    if "id" not in gist:
         print(response_json)
-        print ("Couldn't find an id field. Assume the API request failed - exit.")
+        print("Couldn't find an id field. Assume the API request failed - exit.")
         sys.exit(1)
-    
-    time_created = dateutil.parser.isoparse(gist['created_at'])
-    id = gist['id']
-    age_days = (datetime.now(time_created.tzinfo) - time_created).total_seconds() / 86400.0
+
+    time_created = dateutil.parser.isoparse(gist["created_at"])
+    id = gist["id"]
+    age_days = (
+        datetime.now(time_created.tzinfo) - time_created
+    ).total_seconds() / 86400.0
 
     if age_days > 30:
-        print (f"Gist {id} is {age_days} days old - will delete.")
+        print(f"Gist {id} is {age_days} days old - will delete.")
 
-        r = requests.delete(f'https://api.github.com/gists/{id}', 
-                  headers = headers)
-        print (f'Deleted {id}.')
+        r = requests.delete(f"https://api.github.com/gists/{id}", headers=headers)
+        print(f"Deleted {id}.")
         gists_deleted += 1
-
-
