+++ comment-github-pullrequest
@@ -2,45 +2,50 @@
 """
 Makes a comment on Pull request on Github
 """
-from __future__ import print_function
-from github import Github
-from os.path import expanduser, dirname, abspath, join, exists
-from optparse import OptionParser
-from sys import exit
-import re, sys
-from socket import setdefaulttimeout
-setdefaulttimeout(120)
-SCRIPT_DIR = dirname(abspath(sys.argv[0]))
 
+import os
+import sys
+import argparse
+from socket import setdefaulttimeout
+from github import Github
+
+from Mu2eCI import config
+from Mu2eCI.logger import log
+from Mu2eCI.comment_gh_pr import comment_gh_pr
+
+setdefaulttimeout(120)
+
+parser = argparse.ArgumentParser(
+    description="Process a pull request given a repository and PR number."
+)
+parser.add_argument("--pullrequest", type=int, help="The Pull Request ID.")
+parser.add_argument(
+    "--repository",
+    type=str,
+    help="The GitHub repository. Must be in the format e.g. Mu2e/Offline",
+    choices=config.main["supported_repos"],
+)
+parser.add_argument(
+    "--dry-run",
+    type=bool,
+    default=False,
+    help="Is this a dry run? i.e. don't touch GitHub.",
+)
+
+parser.add_argument(
+    "--report-file", type=str, help="The report file e.g. gh-report.md", default=""
+)
+
+args = parser.parse_args()
 
 
 if __name__ == "__main__":
-    parser = OptionParser(usage="%prog -p|--pullrequest <number> -m|--message <message> [-r|--repository <repo>] [-n|--dry-run]")
-    parser.add_option("-n", "--dry-run",     dest="dryRun",     action="store_true", help="Do not modify Github", default=False)
-    parser.add_option("-r", "--repository",  dest="repository", help="Github Repositoy name e.g. cms-sw/cmssw.",  type=str, default="cms-sw/cmssw")
-    parser.add_option("-p", "--pullrequest", dest="pr",         help="Github Pull Request Number e.g. 10500",     type="int", metavar="N")
-    parser.add_option("-m", "--message",     dest="msg",        help="Message to be added for Github Pull Request", type="str")
-    parser.add_option("-R", "--report-file", dest="report_file",help="Message from the file to be added for Github Pull Request", type="str")
-    opts, args = parser.parse_args()
+    gh = Github(login_or_token=os.environ["GITHUBTOKEN"], retry=3)
 
-
-    if not opts.pr: parser.error("Missing pull request number : -p|--pullrequest <number>")
-    msg = ""
-    if opts.msg: msg = re.sub("@N@","\n",opts.msg)
-    elif opts.report_file: msg = open(opts.report_file).read()
-    else: parser.error("Missing pull request message: -m|--message <message> OR -R|--report-file <file-path>")
-    if opts.dryRun:
-      print("Addeding Comments:",msg)
-    else:
-      repo_dir = join(SCRIPT_DIR,'repos',opts.repository.replace("-","_"))
-      if exists(join(repo_dir,"repo_config.py")): sys.path.insert(0,repo_dir)
-      import repo_config
-      from comment_gh_pr import comment_gh_pr
-      from os import environ
-      gh = Github(login_or_token=environ['GITHUBTOKEN'], retry=3)
-      try:
-          comment_gh_pr(gh, opts.repository, opts.pr, msg)
-      except Exception as e:
-          print("Failed to add comment: ",e)
-          exit(1)
-
+    try:
+        with open(args.report_file, "r") as f:
+            msg = f.read()
+        comment_gh_pr(gh, args.repository, args.pullrequest, msg)
+    except Exception:
+        log.exception("Failed to add comment")
+        sys.exit(1)
