+++ config/main.yaml
@@ -1,0 +1,9 @@
+
+
+bot:
+  username: FNALbuild
+
+supported_repos:
+  - Mu2e/Offline
+
+jenkins_server: https://buildmaster.fnal.gov/buildmaster
