+++ process-pull-request
@@ -1,45 +1,48 @@
 #!/usr/bin/env python
 """
-Returns top commit of a PR (mostly used to comments)
+Process a PR given a repository and PR id number.
 """
-from __future__ import print_function
-from github import Github
-from os.path import expanduser, dirname, abspath, join, exists
-from optparse import OptionParser
-from socket import setdefaulttimeout
-from process_pr import api_rate_limits
-setdefaulttimeout(120)
-import sys
 import os
-SCRIPT_DIR = dirname(abspath(sys.argv[0]))
+import argparse
+from socket import setdefaulttimeout
+
+from github import Github
+from Mu2eCI import config
+from Mu2eCI.common import api_rate_limits
+from Mu2eCI.process_pr import process_pr
+
+setdefaulttimeout(120)
+
+parser = argparse.ArgumentParser(
+    description="Process a pull request given a repository and PR number."
+)
+parser.add_argument("pr_id", type=int, help="The Pull Request ID.")
+parser.add_argument(
+    "repo",
+    type=str,
+    help="The GitHub repository. Must be in the format e.g. Mu2e/Offline",
+    choices=config.main["supported_repos"],
+)
+parser.add_argument(
+    "--dry-run",
+    type=bool,
+    default=False,
+    help="Is this a dry run? i.e. don't touch GitHub.",
+)
+
+args = parser.parse_args()
+
 
 if __name__ == "__main__":
-  parser = OptionParser(usage="%prog <pull-request-id>")
-  parser.add_option("-c", "--commit",     dest="commit",     action="store_true", help="Get last commit of the PR", default=False)
-  parser.add_option("-a", "--all",        dest="all",        action="store_true", help="Get all commits of the PR", default=False)
-  parser.add_option("-n", "--dry-run",    dest="dryRun",     action="store_true", help="Do not modify Github", default=False)
-  parser.add_option("-f", "--force",      dest="force",      action="store_true", help="Force process the issue/PR even if it is ignored.", default=False)
-  parser.add_option("-r", "--repository", dest="repository", help="Github Repositoy name e.g. cms-sw/cmssw.", type=str, default="cms-sw/cmssw")
-  parser.add_option("-u", "--user",      dest="user", help="Github user name used for reporting to PRs and issues", type=str, default="FNALbuild")
-  opts, args = parser.parse_args()
+    prId = args.pr_id
+    gh = Github(login_or_token=os.environ["GITHUBTOKEN"], retry=3)
+    api_rate_limits(gh)
 
-  if len(args) != 1:
-    parser.error("Too many/few arguments")
-  prId = int(args[0])  # Positional argument is "Pull request ID"
+    repo = gh.get_repo(args.repo)
 
-  repo_dir = join(SCRIPT_DIR,'repos',opts.repository.replace("-","_"))
-  if exists(join(repo_dir,"repo_config.py")): sys.path.insert(0,repo_dir)
-  import repo_config
-  gh = Github(login_or_token=os.environ['GITHUBTOKEN'], retry=3)
-  if not opts.commit: api_rate_limits(gh)
-  repo = gh.get_repo(opts.repository)
-  if opts.commit:
-    pr = repo.get_pull(prId)
-    if opts.all:
-      for c in pr.get_commits():
-        print(c.sha)
-    else:
-      print(pr.get_commits().reversed[0].commit.sha)
-  else:
-    from process_pr import process_pr
-    process_pr(repo_config, gh, repo, repo.get_issue(prId), opts.dryRun, cmsbuild_user=opts.user, force=opts.force)
+    process_pr(
+        gh,
+        repo,
+        repo.get_issue(prId),
+        args.dry_run,
+    )
